fixed logic in this_plugin_first in case plugin is somehow not activated

+++ init.php
@@ -50,7 +50,7 @@ function this_plugin_first() {
     $this_plugin = plugin_basename(trim($wp_path_to_this_file));
     $active_plugins = get_option('active_plugins');
     $this_plugin_key = array_search($this_plugin, $active_plugins);
-    if ($this_plugin_key) { // if it's 0 it's the first plugin already, no need to continue
+    if ($this_plugin_key !== false && $this_plugin_key > 0) { // it exists and is not already first
         array_splice($active_plugins, $this_plugin_key, 1);
         array_unshift($active_plugins, $this_plugin);
         update_option('active_plugins', $active_plugins);
